requirements.txt
python-telegram-bot==13.8.1
google-cloud-dialogflow==2.9.1